auth after register

+++ routes/web.php
@@ -16,15 +16,13 @@ Route::group(['namespace' => 'App\Http\Controllers\Blog', 'prefix' => 'blog'], f
     Route::get('/{blog}', \ShowController::class, 'show')->name('blog.show');
     Route::get('/{blog}/edit', \EditController::class, 'edit')->name('blog.edit');
     Route::patch('/{blog}', \UpdateController::class, 'update')->name('blog.update');
-    // Route::delete('/{blog}', \DeleteController::class, 'delete')->name('blog.delete')->middleware('auth:sanctum');
-    Route::delete('/{blog}', \DeleteController::class, 'delete')->name('blog.delete');
+    Route::delete('/{blog}', \DeleteController::class, 'delete')->name('blog.delete')->middleware('auth:sanctum');
+    // Route::delete('/{blog}', \DeleteController::class, 'delete')->name('blog.delete');
     Route::post('/', \StoreController::class, 'store')->name('blog.store');
 });
 
 // X-XSRF-TOKEN GET -> /sanctum/csrf-cookie
-
-
-Route::post('/register', [App\Http\Controllers\Sanctum\RegisterUserController::class, 'store']);
+Route::post('/register', [App\Http\Controllers\Sanctum\RegisterUserController::class, 'store'])->middleware('guest');
 
 
 // Route::group(['namespace' => 'App\Http\Controllers\Sanctum', 'prefix' => 'sanctum'], function (){
